+++ src/Headroom/FileSupport/Dart.hs
@@ -2,19 +2,19 @@
 {-# LANGUAGE QuasiQuotes       #-}
 
 {-|
-Module      : Headroom.FileSupport.Go
-Description : Support for /Go/ source code files
+Module      : Headroom.FileSupport.Dart
+Description : Support for /Dart/ source code files
 Copyright   : (c) 2019-2021 Vaclav Svejcar
 License     : BSD-3-Clause
 Maintainer  : vaclav.svejcar@gmail.com
 Stability   : experimental
 Portability : POSIX
 
-Basic support for /Go language/ source code files. This implementation doesn't
+Basic support for /Dart language/ source code files. This implementation doesn't
 extract any variables or template data.
 -}
 
-module Headroom.FileSupport.Go
+module Headroom.FileSupport.Dart
   ( fileSupport
   )
 where
@@ -26,14 +26,14 @@ import           Headroom.FileSupport.Types          ( FileSupport(..)
                                                      , SyntaxAnalysis(..)
                                                      , defaultFileSupport
                                                      )
-import           Headroom.FileType.Types             ( FileType(Go) )
+import           Headroom.FileType.Types             ( FileType(Dart) )
 
 
 ------------------------------  PUBLIC FUNCTIONS  ------------------------------
 
--- | Implementation of 'FileSupport' for /Go/.
+-- | Implementation of 'FileSupport' for /Dart/.
 fileSupport :: FileSupport
-fileSupport = defaultFileSupport Go syntaxAnalysis
+fileSupport = defaultFileSupport Dart syntaxAnalysis
 
 
 ------------------------------  PRIVATE FUNCTIONS  -----------------------------