+++ test-data/code-samples/dart/sample1.dart
@@ -1,0 +1,11 @@
+
+// This is the
+// header
+
+// this is not header
+
+// this is not the header
+void main() {
+  /* another comment */
+  print('Hello, World!');
+}